requirements.txt
absl-py==1.2.0
cachetools==4.2.4
certifi==2020.6.20
charset-normalizer==2.0.12
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
efficientnet-pytorch==0.7.1
einops==0.4.1
FrEIA @ git+https://github.com/VLL-HD/FrEIA.git@1779d1fba1e21000fda1927b59eeac0a6fcaa284
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
imageio==2.15.0
importlib-metadata==4.8.3
iunets==0.1
joblib==1.1.0
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
memcnn==1.5.1
networkx==2.5.1
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.5.5.64
pathlib2==2.3.7.post1
Pillow==8.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
PyWavelets==1.1.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
SimpleITK==2.1.1.2
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2020.9.3
timm==0.5.4
torch==1.10.1
torchinfo==1.5.4
torchsummary==1.5.1
torchvision==0.11.2
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.10
Werkzeug==2.0.3
yapf==0.32.0
zipp==3.6.0
